requirements.txt
scipy==1.2.1
numpy==1.19.0
matplotlib==3.1.2
opencv_python==4.1.2.30
#torch==1.2.0
#torchvision==0.4.0
tqdm==4.60.0
Pillow==8.2.0
h5py==2.10.0
tensorflow==2.2.0
